+++ preliminaries/qwen2_5vl_lora_sft_SQA3D/multi_runner.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
+#SBATCH --nodes=1
+#SBATCH --ntasks-per-node=1
+#SBATCH --cpus-per-task=96
+#SBATCH --time=0-00:15:00
+#SBATCH --gpus-per-node=h100:4
+#SBATCH --array=1-5
+#SBATCH --output=%N-qwen2_5vl_lora_sft_SQA3Devery24_traineval_resumefromcheckpoint_array-%j.out
 
-TASK_NUMBER=$1
-CLUSTER_NAME=$2
+TASK_NUMBER=${SLURM_ARRAY_TASK_ID}
+CLUSTER_NAME=$1
 
 echo "Task number: $TASK_NUMBER"